Release the latest changes from DEV

+++ database/migrations/2024_01_30_164453_add_reason_field_to_dur_has_datasets_table.php
@@ -20,8 +20,10 @@ return new class () extends Migration {
      */
     public function down(): void
     {
-        Schema::table('dur_has_datasets', function (Blueprint $table) {
-            $table->dropColumn('reason');
-        });
+        if (Schema::hasTable('dur_has_datasets')) {
+            Schema::table('dur_has_datasets', function (Blueprint $table) {
+                $table->dropColumn('reason');
+            });
+        }
     }
 };
